fraud-detection-update-sample/requirements.txt
functions-framework==3.3.0
google-events==0.4.0
google-api-core
protobuf==4.22.1
cloudevents==1.9.0
google-cloud-firestore
earthengine-api==0.1.358
numpy
firebase-admin
scipy
ee
geemap
geopy
protobuf
google-cloud-pubsub
